rocjitsu: Move dbt device kernel tests to a seperate file
